# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "battery life — API для управления заявками расчёта времени работы аккумуляторов",
        "title": "battery life API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/battery_life_type/create-battery_life_type": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый тип аккумулятора. Доступно авторизованному пользователю.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Создать аккумулятор",
                "parameters": [
                    {
                        "description": "Данные аккумулятора",
                        "name": "battery_life_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryTypeJSON"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryTypeJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/battery_life_type/{id}": {
            "get": {
                "description": "Возвращает данные одного аккумулятора.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Получить аккумулятор по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID аккумулятора",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryTypeJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/battery_life_types": {
            "get": {
                "description": "Возвращает все аккумуляторы или фильтрует по названию.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Получить список аккумуляторов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название аккумулятора для поиска",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryTypeListJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/signin": {
            "post": {
                "description": "Принимает логин/пароль, возвращает token и role (moderator|creator).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Вход (получение токена)",
                "parameters": [
                    {
                        "description": "Логин и пароль",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.SignInResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/signout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет токен текущего пользователя из blacklist. 204 No Content.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Выход (удаление токена)",
                "responses": {
                    "204": {
                        "description": "Токен добавлен в blacklist"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/signup": {
            "post": {
                "description": "Регистрирует нового пользователя. Возвращает login и role (moderator|creator).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Логин и пароль",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.SignUpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "RIP2026_internal_app_serializer.BatteryTypeJSON": {
            "type": "object",
            "properties": {
                "capacity_mah": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "photo": {
                    "type": "string"
                },
                "short_description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                },
                "voltage_v": {
                    "type": "number"
                }
            }
        },
        "RIP2026_internal_app_serializer.BatteryTypeListJSON": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryTypeJSON"
                    }
                }
            }
        },
        "RIP2026_internal_app_serializer.SignInRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "RIP2026_internal_app_serializer.SignInResponse": {
            "type": "object",
            "properties": {
                "role": {
                    "description": "moderator | creator",
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "RIP2026_internal_app_serializer.SignUpRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "is_moderator": {
                    "type": "boolean"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "RIP2026_internal_app_serializer.SignUpResponse": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "role": {
                    "description": "moderator | creator",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}